Issue 54: resource hierarchy fixes
All the methods dealing with resources expected the flat
structure, but resources can be nested, so instead of resource_id as
identifier we now use an array of those ids.

Also sorting the results to obtain the consistent response over Hawkular
server restart => no need to update the VCRs over and over again, if
it's the same data up to the ordering.