Fix issue with mixed JSON and binary data in Hawkular websocket message
When invoking "export JDR" operation, Hawkular sends through the
websocket the required JSON response for any operation. Along with the
JSON response, the binary data of the report is also sent. As a
consequence, when parsing the JSON response also the binary data was
being parsed causing an exception.

These changes determine if there is binary data in the response from
Hawkular to parse only the JSON part and separate the binary data
fragment.